+++ CHANGELOG.md
@@ -18,6 +18,7 @@ _Note: Gaps between patch versions are faulty, broken or test releases._
 * Added tests:
   * `bImage`
   * `super/i-block/modules/daemons`
+  * `core/compoent/directives/update-on`
 
 ## v3.0.0-rc.163 (2021-03-19)